MongoDB Installation and Set Up
Setting up MongoDB Installation and using mongoose to create a connection between
the MongoDB and the node.js runtime environment. We utilised the express.js "app.listen()" to bind and listen to the connections on the specified host and port. In otherwords creating a ahttp server object, configuring it to received incoming TCP connections on that specified address.